+++ app/javascript/bundles/Dashboard/components/UserInfoCard/UserInfoCard.jsx
@@ -1,8 +1,10 @@
 import React from "react";
+import axios from 'axios';
 import PropTypes from "prop-types";
 import _ from "lodash";
 import "./user-info-card.scss";
-import {ComposedChart, Bar, Line, XAxis, YAxis, CartesianGrid, Tooltip, Legend} from 'recharts';
+
+import UserInfoChart from "../UserInfoChart";
 
 class UserInfoCard extends React.Component {
   constructor(props) {
@@ -14,12 +16,13 @@ class UserInfoCard extends React.Component {
   }
 
   componentDidMount() {
-    fetch('http://localhost:3000/account/user-info')
-      .then(response => response.json())
-      .then(data => this.setState({ data: data }));
+    axios.get(`/account/user-info`)
+     .then(resp => {
+       this.setState({ data: resp.data.info });
+     })
   }
 
-  render() {
+  render () {
     return (
       <div>
         <UserInfoChart data={this.state.data}/>
@@ -28,21 +31,4 @@ class UserInfoCard extends React.Component {
   }
 }
 
-class UserInfoChart extends React.Component {
-  render () {
-    return (
-      <ComposedChart width={700} height={300} data={this.props.data}
-          margin={{top: 20, right: 30, left: 20, bottom: 5}}>
-        <XAxis dataKey="date" stroke="#000"/>
-        <YAxis orientation="left" stroke="#000"/>
-        <CartesianGrid strokeDasharray="3 3"/>
-        <Tooltip />
-        <Legend />
-        <Bar dataKey='created' barSize={20} fill='#8884d8'/>
-        <Bar dataKey='assigned' barSize={20} fill='#82ca9d'/>
-        <Line type="monotone" dataKey="completed" stroke="#ff7300" />
-      </ComposedChart>
-    );
-  }
-}
  export default UserInfoCard;
