+++ app/javascript/bundles/Dashboard/components/UserInfoChart/index.js
@@ -1,0 +1,3 @@
+import UserInfoChart from './UserInfoChart';
+
+export default UserInfoChart;
